pyproject.toml
[project]
name = "omnidocs"
dynamic = ["version"]
description = "Unified Python toolkit for visual document processing - think Transformers for document AI"
authors = [{name = "adithya-s-kolavi", email = "adithyaskolavi@gmail.com"}]
readme = "README.md"
requires-python = ">=3.10,<3.13"
license = {text = "Apache-2.0"}
keywords = ["document-processing", "pdf", "ocr", "layout-analysis", "vision-language-models"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing",
]
dependencies = [
    "doclayout-yolo>=0.0.4",
    "docling-ibm-models>=3.4.0",
    "hf-transfer>=0.1.0",
    "huggingface-hub>=0.36.0",
    "numpy>=2.2.6",
    "opencv-contrib-python>=4.10.0",
    "pdfplumber>=0.11.9",
    "pillow>=12.1.0",
    "pydantic>=2.12.5",
    "pypdfium2>=5.3.0",
    "requests>=2.32.5",
    "rtree>=1.0.0",
    "timm>=1.0.24",
    "torch>=2.6",
    "torchvision",
    "transformers>=4.57.6,<5.0.0",
]

[project.urls]
Homepage = "https://github.com/adithya-s-k/OmniDocs"
Documentation = "https://adithya-s-k.github.io/OmniDocs/"
Repository = "https://github.com/adithya-s-k/OmniDocs"
Issues = "https://github.com/adithya-s-k/OmniDocs/issues"
Changelog = "https://github.com/adithya-s-k/OmniDocs/releases"

[project.optional-dependencies]
# For pip install omnidocs[extra]
pytorch = [
    "accelerate>=1.0.0",
    "qwen-vl-utils>=0.0.14",
]
vllm = [
    "vllm>=0.6.0",
    "accelerate>=1.0.0",
    "qwen-vl-utils>=0.0.14",
]
mlx = [
    "mlx>=0.20.0",
    "mlx-vlm>=0.1.0",
]
api = [
    "openai>=1.0.0",
]
all = [
    "omnidocs[pytorch,vllm,api]",
]

[dependency-groups]
# For uv sync --group <name>
api = [
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
]
dev = [
    "mike>=2.1.3",
    "mkdocs-gen-files>=0.6.0",
    "mkdocs-git-revision-date-localized-plugin>=1.5.0",
    "mkdocs-literate-nav>=0.6.2",
    "mkdocs-material>=9.7.1",
    "mkdocs-section-index>=0.3.10",
    "mkdocstrings>=1.0.1",
    "mkdocstrings-python>=2.0.1",
    "pytest>=8.0",
    "reportlab>=4.4.9",
    "ruff>=0.14.13",
]
docs = [
    "mkdocs>=1.6.1",
    "mkdocs-gen-files>=0.6.0",
    "mkdocs-git-revision-date-localized-plugin>=1.5.0",
    "mkdocs-literate-nav>=0.6.2",
    "mkdocs-material>=9.7.1",
    "mkdocs-section-index>=0.3.10",
    "mkdocstrings[python]>=1.0.1",
]
mlx = [
    "mlx>=0.30.3",
    "mlx-vlm>=0.3.9",
]
ocr = [
    "easyocr>=1.7.2",
    "paddleocr>=3.4.0",
    "paddlepaddle>=3.3.0",
    "paddlex[ocr-core]>=3.4.0",
    "pytesseract>=0.3.13",
]
pytorch = [
    "accelerate>=1.12.0",
    "qwen-vl-utils>=0.0.14",
    "torch>=2.6",
    "transformers>=4.57.6,<5.0.0",
]
vllm = [
    {include-group = "pytorch"},
    "vllm>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "omnidocs/_version.py"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (requiring network, large files, etc.)",
    "gpu: requires GPU for execution",
    "cpu: CPU-only tests",
    "vllm: VLLM backend tests",
    "pytorch: PyTorch backend tests",
    "mlx: MLX backend tests (Apple Silicon)",
    "api: API backend tests",
    "text_extraction: text extraction task tests",
    "layout_extraction: layout extraction task tests",
    "ocr_extraction: OCR extraction task tests",
    "table_extraction: table extraction task tests",
    "reading_order: reading order prediction task tests",
    "integration: integration tests requiring model inference",
]
filterwarnings = [
    # Ignore deprecation warning from thop (external dependency of doclayout-yolo)
    "ignore:The distutils package is deprecated:DeprecationWarning",
]

[tool.uv]
# Override opencv dependencies from packages that pull in conflicting variants
# easyocr wants opencv-python-headless, paddleocr/paddlex want opencv-contrib-python
# We use opencv-contrib-python as the standard (superset of opencv-python)
override-dependencies = [
    "opencv-contrib-python>=4.10.0",
    # Replace headless/standard with opencv-contrib-python
    "opencv-python-headless>=4.10.0 ; sys_platform == 'never'",
    "opencv-python>=4.10.0 ; sys_platform == 'never'",
]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"]
